PairTasks

+++ IE119-Lab2/LcdDriver/Crystalfontz128x128_ST7735.h
@@ -115,7 +115,7 @@ extern void Crystalfontz128x128_LineDrawH(const Graphics_Display *pDisplay,
                                                int16_t lX1,
                                                int16_t lX2,
                                                int16_t lY,
-                                               uint16_t ulValue);
+                                               uint32_t ulValue);
 
 
 #endif /* __CRYSTALFONTZLCD_H__ */
