+++ IE119-Lab2/LcdDriver/HAL_MSP_EXP432P401R_Crystalfontz128x128_ST7735.c
@@ -107,7 +107,7 @@ void HAL_LCD_writeCommand(uint8_t command)
 // interface to the LCD display.
 //
 //*****************************************************************************
-void HAL_LCD_writeData(uint8_t data)
+void HAL_LCD_writeData(uint32_t data)
 {
     // USCI_B0 Busy? //
     while (UCB0STATW & UCBUSY);
