+++ IE119-Lab2/LcdDriver/HAL_MSP_EXP432P401R_Crystalfontz128x128_ST7735.h
@@ -79,7 +79,7 @@
 //
 //*****************************************************************************
 extern void HAL_LCD_writeCommand(uint8_t command);
-extern void HAL_LCD_writeData(uint8_t data);
+extern void HAL_LCD_writeData(uint32_t data);
 extern void HAL_LCD_PortInit(void);
 extern void HAL_LCD_SpiInit(void);
